Bugfix to the homing missile algorithm.
I had originally written it so that the missile performed scans in the direction it
was moving, but this made the missile too easy to evade, so now the missile points
the scanner in the direction in the last known direction of the enemy (utilizing
self.orient for that). Probably too hard to evade now, but w/e.