Added an output that shows the current view range. Useful so you can get an idea of scale.
The top output is the number of 'swaps', that is the number of times the data has been redrawn. If this stalls, the universe has stopped sending outputs.

The bottom is the view range currently displayed. The first pair is the X values, and the second pair is the Y values.