Proof of concept motion control!
The visualization client now supports the complete client negotiation gauntlet (pick a ship, or a class, rename a ship, disconnect), as well as rudimentry thrust-based motion of the chosen smarty. It uses PHYSPROPS messages that are proxied through the ship object, and it seems to be able to so about thirty updates per second. Physics ticks were at a steady 0.1ms with only the single smartie in the universe, and visdata updates were set at 20 FPS.

Some modifications around the Python to add support for this.