Fixed a bug in the Unix codepaths in the universe sim regarding time deltas.
Also added logic to ignore collisions with a combined energy less than one nanoJoule.
As the comment says, this is to ignore spurious collisions that will occur briefly after objects
collide as the physics tick them out of each other's bounding boxes. Because the objects will be
moving apart, the collision energy will be zero (or very close, within the a double's sigfigs,
so to about 1e-15), and thus we can ignore them.