Added a cutoff for the number of collision rounds that consider multi-collision instants.
This restores interactivity, while still allowing the simulation to take advantage of the the increased accuracy of the multi-collision instants for some period of time.

With a rack of 15 rows, on a x86-debug build, run with debugging, it took about two seconds to compute the 1100 rounds of collisions.
On x64-release run without debugging, there was barely a hiccup (about 30ms to compute the event).